src/requirements.txt
attrs==23.1.0
aws-lambda-typing==2.17.0
aws-requests-auth==0.4.3
aws-xray-sdk==2.12.0
boto3==1.26.114
botocore==1.29.114
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
dataclasses-json==0.5.7
idna==3.4
jmespath==1.0.1
jsonschema==4.17.3
jsonschema-spec==0.1.4
lazy-object-proxy==1.9.0
marshmallow==3.19.0
marshmallow-dataclass==8.5.3
marshmallow-enum==1.5.1
marshmallow-jsonschema==0.9.0
mypy-extensions==1.0.0
openapi-schema-validator==0.4.4
openapi-spec-validator==0.5.6
orjson==3.8.10
packaging==23.1
pathable==0.4.3
pycparser==2.21
pynamodb==5.4.1
pynamodb-attributes==0.4.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pyyaml==6.0
requests==2.28.2
rfc3339-validator==0.1.4
s3transfer==0.6.0
six==1.16.0
typing-extensions==4.5.0
typing-inspect==0.7.1
urllib3==1.26.15
waiting==1.4.1
wrapt==1.15.0

# The following packages are considered to be unsafe in a requirements file:
pip==23.1
